Rework project fixing bugs. (#4)
Fixed what seems most of the existent bugs, the protocol is working
as expected now. Did some changes removing the structure that 
resembles a KV storage, and starting to create an API that is much 
like a Send/Listen only.

Some cleanup still needed, the method comments also need to be
rewritten to catch up. Some of the files should be removed, maybe
create a new project for utilities only.

The next steps should be creating tests to ensure that the protocol
works when using the `generic` features, validating the conflict
relationship and that the messages are still delivered in the same
order for all peers.

Some work also should be done to create a transport for sending
messages that do not need a reliable transport. At this time, all
messages are sent using the atomic broadcast implementation and
this should be avoided to truly validate the project.