Listen API.
Create an API for listening for commits applied to a peer WAL. Using
this API is easier to implement tests and know when a peer stop
receiving commands, to there is no need to rely on sleeping a arbitrary
amount of seconds before reading the command history.

In the future, the log and storage abstraction can be removed, since
they are not used by the protocol. After removing we will have something
more close to a communication primitive, where we only send messages to
processes and listen received commands.

Also applied some changes so the CI could keep up, I think that GitHub
actions are too limited unfortunately, so now only the tests on the
`test` folder will be executed on the GitHub actions and the complete
test suite will be executed on CircleCI.