Generic Multicast 0 TLA specification
This spec verify the first generic multicast algorithm, that works in a
failure free environment. This is just a toy algorithm and we started
from here, so we can later just increase the needed steps from one
algorithm to another.

Just in this specification we already had some interesting results, that
could cause the protocol to no satisfy some properties. The first one is
on step `assignSeqNumber`, if the decided timestamp is greater than the
process local clock, we leap to the decided value. This could lead to
some inconsistency, when messages contains the same timestamp they are
sorted by the unique identifier, but if messages are too delayed and
they are supposed to delay with the message already delivered, both
messages will contain the same final timestamp but could be delivered
out of order. To solve this problem we also verify if the process clock
needs to be incremented on this step, if there is a message that
conflict with the current message, the clock will leap and increase, if
there is no conflicting message, the clock only leap.

Another change is on the `doDeliver` step. The definition presented on
the pseudo-code has some details that can lead to delivering messages
out of order. If there two messages with both timestamps, the unique
identifier will be used to break the tie, but if the message with the
smallest identifier is not on the `Delivering` set, the message with
higher identifier will be delivered first. When explaining how the
deliver method will work, the identifier is not used to break ties. To
solve this problem, instead of using a different function to verify the
messages ready to deliver, we use only a single expression that is able
to verify if the message is the smallest amongst all sets, already
verifying both timestamp, conflict relation and the unique identifier.
When ready to delivered, the commuting messages are also added to the
set to be delivered, so the generic property remains working.

In the step `computeSeqNumber`, the received messages and timestamps are
implicitly stored somewhere, so the coordinator can decide later which
is the highest received timestamp. The pseudo-code abstract this, but
when writing the specification a structure called `Votes` is added, this
does not change the protocol itself, the only difference is that now is
explicit the need to keep track of the received timestamps before
deciding the final timestamp.

The last change was adding one more property to the protocol. Using only
the _Partial Order_ we want that the protocol sort messages in a certain
way, but we never tell the algorithm that it _should_ sort messages.
With only the _Partial Order_ definition, messages delivered at the same
instant are accepted, but we know that if messages conflict they should
be delivered in different instant. To avoid this situation, another
property is added to tell the protocol it must sort conflicting
messages.

This property is called _Collision_ (perhaps is an overloaded term?). If
a process `p` deliver both messages `m` and `n`, if `m` and `n` are
different, `p` is both in `m.d` and `n.d` and `m` conflicts with `n`,
then `p` delivered `m` in a different instant than when `p` delivered
`n`. Using this property, now the protocol knows that messages that
conflict are supposed to be sorted.

Those are the major findings when validating the protocol with TLA+.
With this result we are able to see that only reasoning about a protocol
and possible states is something really hard, since so many edge cases
can be created. Specially the need for another property, when only
reading the properties, that conflicting messages can be delivered
after another by a process if, and only if other process also delivered
the message after. Only reading this statement is not easy to see that
the protocol need one more property to secure the ordering of messages.

---

Since not everything is happiness, there are some tradeoffs in this
specification. The most notable is the network abstraction, the protocol
definition is a genuine protocol, thus only processes involved in the
destination and source need to process the message so the protocol can
proceed. In the specification we are using all processes when exchaning
messages between processes, in this approach the number of generated
stated will be smaller. To strictly specify the protocol, when
exchanging messages only the processes on the destination of the message
is needed, but since TLC will verify __every__ state, if we used a SUBSET of
the available Processes all valid subsets will be verified, leading to a
huge number of states. When reasoning about, using all Processes or
using only a subset should not affect the protocol behaviour nor fail
when verifying the properties, but then again, the problems we found
were only found because we verified with TLC and when only reasoning
about the protocol, it did not seem to have any problems.